some images for testing vision algorithms
These were captured with the Axis2100 camera. They are of the atrium
with Andrea in different locations holding an LED flashlight. The images
look out of focus because the lens on the camera is a short depth of
field lens, so doesn't work well for this application. We should have a
new camera soon.